Include `--listfullpaths` in `makeprg`
Before this change, running `:make` on a file with errors when not
in the same directory as the file would open a buffer in the cwd.
This change fixes this: the compiler outputs the full path to the
file, so a buffer is always opened for the correct file.

`--listfullpaths` was used instead of `--listFullPaths` to be
consistent with `autoload/nim.vim`.